put threaded workers into their own class
